move responsive table to platform
