+++ src/platform/forms-system/src/js/components/ResponsiveTable.jsx
